requirements.txt
python-telegram-bot==20.7
google-api-python-client
google-auth
google-auth-oauthlib
openai
python-dotenv

fastapi
uvicorn
